Model: Shift currentDate functionality to AddressBook, update ModelManager to be able to get schedules and regex fix for timing
