+++ src/main/java/seedu/address/model/ReadOnlyAddressBook.java
@@ -2,6 +2,7 @@ package seedu.address.model;
 
 import javafx.collections.ObservableList;
 import seedu.address.model.event.ReadOnlyEvent;
+import seedu.address.model.event.timeslot.Date;
 import seedu.address.model.person.ReadOnlyPerson;
 import seedu.address.model.relationship.Relationship;
 import seedu.address.model.tag.Tag;
@@ -23,14 +24,24 @@ public interface ReadOnlyAddressBook {
      */
     ObservableList<Tag> getTagList();
 
+    //@@author A0162268B
     /**
      * Returns an unmodifiable view of the events list.
      */
     ObservableList<ReadOnlyEvent> getEventList();
+
+    /**
+     * Returns an unmodifiable view of the schedule.
+     */
+    ObservableList<ReadOnlyEvent> getSchedule(Date currentDate);
+
+    //@@author
+
     /**
      * Returns an unmodifiable view of the tags list.
      * This list will not contain any duplicate tags.
      */
+
     ObservableList<Relationship> getRelList();
 
     /**
@@ -38,7 +49,5 @@ public interface ReadOnlyAddressBook {
      */
     ReadOnlyEvent getLastChangedEvent();
 
-
-
-
+    Date getCurrentDate();
 }
